GridGraphs.jl

Welcome to the documentation for GridGraphs, a package made for efficient analysis of rectangular grid graphs.

Index

GridGraphs.FullGridGraph

GridGraphs.GridGraph

GridGraphs.MoveCost

GridGraphs.MoveDirection

GridGraphs.MoveType

GridGraphs.ShortestPathTree

GridGraphs.SparseGridGraph

Graphs.weights

GridGraphs.coord_to_index

GridGraphs.get_path

GridGraphs.grid_bellman_ford

GridGraphs.grid_bellman_ford

GridGraphs.grid_dijkstra

GridGraphs.grid_dijkstra

GridGraphs.grid_topological_sort

GridGraphs.grid_topological_sort

GridGraphs.has_negative_weights

GridGraphs.height

GridGraphs.index_to_coord

GridGraphs.inneighbors_coord

GridGraphs.outneighbors_coord

GridGraphs.path_to_matrix

GridGraphs.vertex_weight

GridGraphs.vertex_weight_coord

GridGraphs.width

Docstrings

GridGraphs.FullGridGraph — Type

FullGridGraph

Concrete subtype of GridGraph for which all vertices are active.

GridGraphs.GridGraph — Type

GridGraph{T<:Integer,R<:Real,W<:AbstractMatrix{R},A<:AbstractMatrix{Bool},mt,md,mc}

Abstract supertype for graphs defined by a grid of vertices.

Required fields

weights::W: vertex weights matrix. The weight of an edge is a simple function of a few vertex weights, depending on the type of grid.

active::A: vertex activity matrix. All the vertices on the grid exist, but only some are active (i.e. can have edges with their neighbors).

Type parameters

T: type of vertex indices

R: type of edge weights

W: type of vertex weights matrix

A: type of vertex activity matrix

mt::MoveType

md::MoveDirection

mc::MoveCost

GridGraphs.MoveCost — Type

MoveCost

Enum type with possible values direct (the edge weight is the weight of the destination) or corner (the edge weight follows Pythagoras' theorem).

GridGraphs.MoveDirection — Type

MoveDirection

Enum type with possible values cyclic (all possible directions) or acyclic (only down and right).

GridGraphs.MoveType — Type

MoveType

Enum type with possible values rook (4 neighbors) or queen (8 neighbors).

GridGraphs.ShortestPathTree — Type

ShortestPathTree{T<:Integer,R<:Real}

Storage for the result of a single-source shortest paths query with source s.

Fields

parents::Vector{T}: the parent of each vertex v in a shortest s -> v path.

dists::Vector{R}: the distance of each vertex v from s.

GridGraphs.SparseGridGraph — Type

FullGridGraph

Concrete subtype of GridGraph for which only some vertices are active.

Graphs.weights — Method

Graphs.weights(g)

Compute a sparse matrix of edge weights based on the vertex weights.

GridGraphs.coord_to_index — Method

coord_to_index(g, i, j)

Convert a grid coordinate tuple (i,j) into the index v of the associated vertex.

GridGraphs.get_path — Method

get_path(spt::ShortestPathTree, s, d)

Reconstruct the shortest s -> d path from a ShortestPathTree with source s.

GridGraphs.grid_bellman_ford — Method

grid_bellman_ford(g, s, d)

Apply grid_bellman_ford(g, s) and retrieve the shortest path from s to d.

GridGraphs.grid_bellman_ford — Method

grid_bellman_ford(g, s)

Apply the Bellman-Ford algorithm on an GridGraph g, and return a ShortestPathTree with source s.

GridGraphs.grid_dijkstra — Method

grid_dijkstra(g, s, d)

Apply grid_dijkstra(g, s) and retrieve the shortest path from s to d.

GridGraphs.grid_dijkstra — Method

grid_dijkstra(g, s)

Apply Dijkstra's algorithm on an GridGraph g, and return a ShortestPathTree with source s.

Uses a DataStructures.BinaryHeap internally instead of a DataStructures.PriorityQueue.

GridGraphs.grid_topological_sort — Method

grid_topological_sort(g, s, d)

Apply grid_topological_sort(g, s) and retrieve the shortest path from s to d.

GridGraphs.grid_topological_sort — Method

grid_topological_sort(g, s)

Apply the topological sort on an acyclic GridGraph g, and return a ShortestPathTree with source s.

Assumes vertex indices correspond to topological ranks.

GridGraphs.has_negative_weights — Method

has_negative_weights(g)

Check whether there are any negative weights.

GridGraphs.height — Method

height(g)

Compute the height of the grid (number of rows).

GridGraphs.index_to_coord — Method

index_to_coord(g, v)

Convert a vertex index v into the tuple (i,j) of associated grid coordinates.

GridGraphs.inneighbors_coord — Function

inneighbors_coord(g, i, j)

Return the inneighbors of (i, j) listed in ascending index order.

GridGraphs.outneighbors_coord — Function

outneighbors_coord(g, i, j)

Return the outneighbors of (i, j) listed in ascending index order.

GridGraphs.path_to_matrix — Method

path_to_matrix(g::GridGraph, path::Vector{<:Integer})

Store the shortest s -> d path in g as an integer matrix of size height(g) * width(g), where entry (i,j) counts the number of visits to the associated vertex.

GridGraphs.vertex_weight — Method

vertex_weight(g, v)

Retrieve the vertex weight associated with index v.

GridGraphs.vertex_weight_coord — Method

vertex_weight_coord(g, i, j)

Retrieve the vertex weight associated with coordinates (i, j).

GridGraphs.width — Method

width(g)

Compute the width of the grid (number of columns).