GridGraphs.jl

Welcome to the documentation for GridGraphs, a package made for efficient analysis of rectangular grid graphs.

Index

GridGraphs.AbstractGridGraph

GridGraphs.AcyclicGridGraph

GridGraphs.GridGraph

GridGraphs.ShortestPathTree

GridGraphs.SparseGridGraph

Graphs.weights

GridGraphs.get_path

GridGraphs.get_weight

GridGraphs.get_weight

GridGraphs.grid_bellman_ford

GridGraphs.grid_bellman_ford

GridGraphs.grid_dijkstra

GridGraphs.grid_dijkstra

GridGraphs.grid_neighbors

GridGraphs.grid_topological_sort

GridGraphs.grid_topological_sort

GridGraphs.has_negative_weights

GridGraphs.height

GridGraphs.is_acyclic

GridGraphs.node_coord

GridGraphs.node_index

GridGraphs.path_to_matrix

GridGraphs.width

Docstrings

GridGraphs.AbstractGridGraph — Type

AbstractGridGraph{T<:Integer,R}

Abstract supertype for grid graphs with vertices of type T and weights of type R.

All subtypes must have a field weights::Matrix{R}, whose size gives the size of the grid and whose entries correspond to vertex weights. The weight of an edge (s,d) is then defined as the weight of the vertex d.

To implement a concrete subtype G <: AbstractGridGraph, the following methods need to be defined (see the Graphs.jl docs):

Graphs.ne(g::G)

Graphs.has_edge(g::G, s, d)

Graphs.outneighbors(g::G, s)

Graphs.inneighbors(g::G, d)

GridGraphs.AcyclicGridGraph — Type

AcyclicGridGraph{T<:Integer,R}

Concrete subtype of AbstractGridGraph, in which we can move from a cell (i,j) to its bottom, right and bottom right neighbors only. This means the graph is acyclic.

Fields

weights::Matrix{R}: grid of vertex weights

GridGraphs.GridGraph — Type

GridGraph{T<:Integer,R}

Concrete subtype of AbstractGridGraph, in which we can move from a cell (i,j) to any of its 8 nearest neighbors (lateral, vertical and diagonal).

Fields

weights::Matrix{R}: grid of vertex weights

GridGraphs.ShortestPathTree — Type

ShortestPathTree{T<:Integer,R<:Real}

Storage for the result of a single-source shortest paths query with source s.

Fields

parents::Vector{T}: the parent of each vertex v in a shortest s -> v path.

dists::Vector{R}: the distance of each vertex v from s.

GridGraphs.SparseGridGraph — Type

SparseGridGraph{T<:Integer,R}

Analogue of GridGraph in which only some vertices are active.

For simplicity, we consider that all nodes of the grid exist, but only some are usable (i.e. linked to their neighbors), so that masked vertices are alone in their connected component.

Fields

weights::Matrix{R}: grid of vertex weights

active::Matrix{Bool}: grid of boolean values, in which true indicates that a vertex is active (i.e. can be used)

ne::Int: number of edges, precomputed at construction

Note that for simplicity, inactive vertices still belong to the graph, but they are isolated from their neighbors.

Graphs.weights — Method

Graphs.weights(g)

Compute a sparse matrix of edge weights based on the vertex weights.

GridGraphs.get_path — Method

get_path(spt::ShortestPathTree, s, d)

Reconstruct the shortest s -> d path from a ShortestPathTree with source s.

GridGraphs.get_weight — Method

get_weight(g, i, j)

Retrieve the vertex weight associated with coordinates (i,j).

GridGraphs.get_weight — Method

get_weight(g, v)

Retrieve the vertex weight associated with index v.

GridGraphs.grid_bellman_ford — Method

grid_bellman_ford(g, s, d)

Apply grid_bellman_ford(g, s) and retrieve the shortest path from s to d.

GridGraphs.grid_bellman_ford — Method

grid_bellman_ford(g, s)

Apply the Bellman-Ford algorithm on an AbstractGridGraph g, and return a ShortestPathTree with source s.

GridGraphs.grid_dijkstra — Method

grid_dijkstra(g, s, d)

Apply grid_dijkstra(g, s) and retrieve the shortest path from s to d.

GridGraphs.grid_dijkstra — Method

grid_dijkstra(g, s)

Apply Dijkstra's algorithm on an AbstractGridGraph g, and return a ShortestPathTree with source s.

Uses a DataStructures.BinaryHeap internally instead of a DataStructures.PriorityQueue.

GridGraphs.grid_neighbors — Method

grid_neighbors(g, i, j)

Return an iterator of grid neighbors listed in ascending index order.

GridGraphs.grid_topological_sort — Method

grid_topological_sort(g, s, d)

Apply grid_topological_sort(g, s) and retrieve the shortest path from s to d.

GridGraphs.grid_topological_sort — Method

grid_topological_sort(g, s)

Apply the topological sort on an acyclic AbstractGridGraph g, and return a ShortestPathTree with source s.

Assumes vertex indices correspond to topological ranks.

GridGraphs.has_negative_weights — Method

has_negative_weights(g)

Check whether the graph g has any negative weight.

GridGraphs.height — Method

height(g)

Compute the height of the grid (number of rows).

GridGraphs.is_acyclic — Method

is_acyclic(g)

Check whether g contains cycles.

GridGraphs.node_coord — Method

node_coord(g, v)

Convert a vertex index v into the associate tuple (i,j) of grid coordinates.

GridGraphs.node_index — Method

node_index(g, i, j)

Convert a grid coordinate tuple (i,j) into the index v of the associated vertex.

GridGraphs.path_to_matrix — Method

path_to_matrix(g::AbstractGridGraph, path::Vector{<:Integer})

Store the shortest s -> d path in g as an integer matrix of size height(g) * width(g), where entry (i,j) counts the number of visits to the associated vertex.

GridGraphs.width — Method

width(g)

Compute the width of the grid (number of columns).