GridGraphs.jl

API reference

Public

GridGraphs.GridGraphs — Module

GridGraphs

A package for graphs defined by a rectangular grid of vertices.

GridGraphs.GridGraph — Type

GridGraph

Directed graph defined by a rectangular grid of vertices.

Constructors

GridGraph(vertex_weights; torus=false)

vertex_weights::AbstractMatrix: vertex weight matrix used to define edge weights.

torus::Bool: whether the grid wraps around at the borders.

GridGraph(nrows, ncolumns; torus=false)

nrows and ncolumns: size of the grid (assumes homogeneous vertex weights)

GridGraphs.coord_to_index — Method

coord_to_index(g, i, j)

Convert a grid coordinate tuple (i,j) into the index v of the associated vertex.

GridGraphs.edge_weight — Method

edge_weight(g, s, d)

Compute the weight of the edge from s to d.

Only use this on edges that are guaranteed to exist.

GridGraphs.height — Method

height(g)

Compute the height of the grid (number of rows).

GridGraphs.index_to_coord — Method

index_to_coord(g, v)

Convert a vertex index v into the tuple (i,j) of associated grid coordinates.

GridGraphs.vertex_weight — Method

vertex_weight(g, v)

Retrieve the vertex weight associated with index v.

GridGraphs.width — Method

width(g)

Compute the width of the grid (number of columns).

Private

GridGraphs.GridDirection — Type

GridDirection

Enum type for the 4 possible move directions on a 2-dimensional grid with square cells: west, north, south, east.

GridGraphs.get_direction — Method

get_direction(Δi, Δj)

Translate a couple of grid steps in {±1,0} into a GridDirection.

GridGraphs.get_tuple — Method

get_tuple(dir)

Translate a GridDirection into a couple of grid steps in {±1,0}.

GridGraphs.is_torus — Method

is_torus(g)

Check whether the grid is a torus.

GridGraphs.vertex_weights — Method

vertex_weights(g)

Retrieve the vertex weights matrix.