GridGraphs.jl

GridGraphs.GridGraphs — Module

GridGraphs

A package for graphs defined by a rectangular grid of vertices.

GitHub repo: https://github.com/gdalle/GridGraphs.jl

API reference

Types

GridGraphs.GridDirection — Type

GridDirection

Enum type for the 9 possible move directions on a 2-dimensional grid with square cells: northwest, west, southwest, north, center, south, northeast, east, southeast.

Various subsets of these directions are defined as constants. They are based on an analogy with the game of chess:

QUEEN_DIRECTIONS

ROOK_DIRECTIONS

QUEEN_DIRECTIONS_PLUS_CENTER

ROOK_DIRECTIONS_PLUS_CENTER

QUEEN_DIRECTIONS_ACYCLIC

ROOK_DIRECTIONS_ACYCLIC

Acyclic direction sets give rise to an acyclic graph because they are contained in {south, east, southeast}.

GridGraphs.GridGraph — Type

GridGraph{
    R<:Real,
    W<:AbstractMatrix{R},
    A<:AbstractMatrix{Bool},
    D<:NTuple{<:Any,GridDirection}
}

Graph defined by a rectangular grid of vertices.

Fields

vertex_weights::W: Vertex weight matrix used to define edge weights.

vertex_activities::A: Vertex activity matrix. All the vertices on the grid exist, but only active vertices can have edges (inactive vertices are isolated, they correspond to obstacles).

directions::D: Set of legal directions used to define edges.

nb_corners_for_diag::Int: Number of active corner vertices necessary for a diagonal edge to exist. Can take the values 0, 1 or 2.

pythagoras_cost_for_diag::Bool: Whether the weight of a diagonal edge is computed using the shortest of the active corner paths (true) or just the weight of the arrival vertex (false).

Constructors

There is a user-friendly constructor with the following default values:

GridGraph(
    vertex_weights;
    vertex_activities=Trues(size(weights)),
    directions=ROOK_DIRECTIONS,
    nb_corners_for_diag=0,
    pythagoras_cost_for_diag=false
)

Access

GridGraphs.get_tuple — Function

get_tuple(dir)

Translate a GridDirection into a couple of grid steps in {±1,0}.

GridGraphs.get_direction — Function

get_direction(Δi, Δj)

Translate a couple of grid steps in {±1,0} into a GridDirection.

GridGraphs.height — Function

height(g)

Compute the height of the grid (number of rows).

GridGraphs.width — Function

width(g)

Compute the width of the grid (number of columns).

GridGraphs.index_to_coord — Function

index_to_coord(g, v)

Convert a vertex index v into the tuple (i,j) of associated grid coordinates.

GridGraphs.coord_to_index — Function

coord_to_index(g, i, j)

Convert a grid coordinate tuple (i,j) into the index v of the associated vertex.

GridGraphs.vertex_active — Function

vertex_active(g, v)

Check if vertex v is active.

GridGraphs.has_direction — Function

has_direction(g, dir)

Check if direction dir is a valid edge direction.

GridGraphs.vertex_weight — Function

vertex_weight(g, v)

Retrieve the vertex weight associated with index v.

GridGraphs.edge_weight — Function

edge_weight(g, s, d)

Compute the weight of the edge from s to d.

Only use this on edges that are guaranteed to exist.

Shortest paths

GridGraphs.get_path — Function

get_path(spt::ShortestPathTree, s, d)

Reconstruct the shortest s -> d path from a ShortestPathTree with source s.

GridGraphs.path_to_matrix — Function

path_to_matrix(g::GridGraph, path::Vector{<:Integer})

Store the shortest s -> d path in g as an integer matrix of size height(g) * width(g), where entry (i,j) counts the number of visits to the associated vertex.

GridGraphs.grid_dijkstra — Function

grid_dijkstra(g, s)

Apply Dijkstra's algorithm on an GridGraph g, and return a ShortestPathTree with source s.

Uses a DataStructures.BinaryHeap internally instead of a DataStructures.PriorityQueue. Compatible with ForwardDiff.

grid_dijkstra(g, s, d)

Apply grid_dijkstra(g, s) and retrieve the shortest path from s to d.

GridGraphs.grid_bellman_ford — Function

grid_bellman_ford(g, s)

Apply the Bellman-Ford algorithm on an GridGraph g, and return a ShortestPathTree with source s.

Compatible with ForwardDiff.

grid_bellman_ford(g, s, d)

Apply grid_bellman_ford(g, s) and retrieve the shortest path from s to d.

GridGraphs.grid_topological_sort — Function

grid_topological_sort(g, s)

Apply the topological sort on an acyclic GridGraph g, and return a ShortestPathTree with source s.

Assumes vertex indices correspond to topological ranks. Compatible with ForwardDiff.

grid_topological_sort(g, s, d)

Apply grid_topological_sort(g, s) and retrieve the shortest path from s to d.

GridGraphs.ShortestPathTree — Type

ShortestPathTree{R<:Real}

Storage for the result of a single-source shortest paths query with source s.

Fields

parents::Vector{Int}: the parent of each vertex v in a shortest s -> v path.

dists::Vector{R}: the distance of each vertex v from s.

Index

GridGraphs.GridGraphs

GridGraphs.GridDirection

GridGraphs.GridGraph

GridGraphs.ShortestPathTree

GridGraphs.coord_to_index

GridGraphs.edge_weight

GridGraphs.get_direction

GridGraphs.get_path

GridGraphs.get_tuple

GridGraphs.grid_bellman_ford

GridGraphs.grid_dijkstra

GridGraphs.grid_topological_sort

GridGraphs.has_direction

GridGraphs.height

GridGraphs.index_to_coord

GridGraphs.path_to_matrix

GridGraphs.vertex_active

GridGraphs.vertex_weight

GridGraphs.width